Rewrite IntervalSet data format to match Coordinates.elements
Now use a list of dict with 'name' and 'interval' key to define the elements
of a temporal dimension (by convention).

The 'interval' key should contain a list of at least one two-element list
(or tuple, though tuples are awkward for yaml serialisation) which defines
the interval extent, as previously.